+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 * reorg of src/
 * added icons (default and platform-specific); they are empty bits
 * tweaked platform PWA manifests
-* provided more instructional comments in head element of index.html
+* provided more instructional comments in head element of html pages
 * added webpack-dashboard support
 * added webpack-bundle-analyzer support
 * added google lighthouse CLI support